+++ packages/hooks/src/services/painter/basePaint.ts
@@ -1,4 +1,4 @@
-import { OrderSide } from "@orderly.network/types";
+import { MarginMode } from "@orderly.network/types";
 import { PosterPainter } from "./painter";
 
 export type posterDataSource = {
@@ -11,7 +11,7 @@ export type posterDataSource = {
     brokerName?: string;
     // side: OrderSide;
     side: "LONG" | "SHORT";
-    marginMode?: "isolated" | "cross";
+    marginMode?: MarginMode;
     /**
      * The leverage of the position
      */