+++ packages/ui-share/src/types/types.tsx
@@ -1,3 +1,5 @@
+import { MarginMode } from "@orderly.network/types";
+
 export type layoutInfo = {
   width?: number;
   height?: number;
@@ -94,7 +96,7 @@ export type ShareOptions =
 export type ShareEntity = {
   symbol: string;
   side: "LONG" | "SHORT";
-  marginMode?: "isolated" | "cross";
+  marginMode?: MarginMode;
   pnl?: number;
   roi?: number;
   openPrice?: number;