+++ packages/net/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @orderly.network/net
 
+## 2.11.1
+
+### Patch Changes
+
+- publish
+- 2bb6924: publish
+
+## 2.11.1-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 2.11.0
 
 ### Minor Changes