+++ packages/web3-provider-ethers/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @orderly.network/web3-provider-ethers
 
+## 2.11.1
+
+### Patch Changes
+
+- publish
+- 2bb6924: publish
+- Updated dependencies
+- Updated dependencies [2bb6924]
+  - @orderly.network/default-evm-adapter@2.11.1
+  - @orderly.network/types@2.11.1
+
+## 2.11.1-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/default-evm-adapter@2.11.1-alpha.0
+  - @orderly.network/types@2.11.1-alpha.0
+
 ## 2.11.0
 
 ### Minor Changes